Moved badger type into lowest component
